fix for bithound
