+++ .bithoundrc
@@ -4,7 +4,6 @@
     ],
     "critics": {
         "lint": {
-            "configLocation": ".eslintrc-bithound.js",
             "engine": "eslint"
         }
     },
